test: service 테스트 분리
